backend/requirements.txt
agno>=2.5.7
sqlalchemy>=2.0.0
fastapi>=0.115.0
sse-starlette>=3.2.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.0.0
httpx>=0.27.0
trafilatura>=2.0.0
pypdf>=4.0.0
python-docx>=1.1.0
python-multipart>=0.0.20
python-dotenv>=1.0.0
openai>=1.56.0
groq>=0.9.0
tavily-python>=0.3.0
tenacity>=8.2.0
anthropic>=0.34.0
debugpy>=1.8.1

# Test dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Observability
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
openinference-instrumentation-agno>=0.1.0